ETL_stack/web_app/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2-binary
requests
